banco_dados/requirements.txt
mysql-connector==2.2.9
